backend/pyproject.toml
# SPDX-License-Identifier: Apache-2.0
# Copyright (c) 2026 BuildWorks.AI
# EUCORA Control Plane - Python Project Configuration
# BuildWorks.AI Compliance: pyproject.toml is MANDATORY (requirements.txt is FORBIDDEN)

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eucora-control-plane"
version = "1.0.0"
description = "Enterprise Endpoint Application Packaging & Deployment Factory - Control Plane"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "BuildWorks.AI", email = "info@buildworks.ai" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
]

# Core Dependencies (Base Requirements)
dependencies = [
    # Core Framework
    "Django>=5.1.12,<5.2",  # Updated from 5.0.14 to fix CVE-2025-48432, CVE-2025-57833
    "djangorestframework~=3.15.0",
    # Database
    "psycopg2-binary~=2.9.9",
    # Caching & Sessions
    "django-redis~=5.4.0",
    "redis~=5.0.1",
    # API Documentation
    "drf-spectacular~=0.27.2",
    # CORS
    "django-cors-headers~=4.3.1",
    # Environment Configuration
    "python-decouple~=3.8",
    # Structured Logging
    "python-json-logger~=2.0.7",
    # Object Storage (MinIO)
    "minio~=7.2.3",
    # HTTP Requests
    "requests>=2.32.4,<3.0",  # Updated from 2.31.0 to fix CVE-2024-35195, CVE-2024-47081
    # Date/Time Utilities
    "python-dateutil~=2.8.2",
    # Utilities
    "pytz~=2024.1",
    # AI/LLM Provider SDKs
    "openai>=2.0.0",
    "anthropic>=0.34.0",
    "groq>=0.9.0",
    # Async utilities for AI calls
    "nest-asyncio~=1.6.0",
    # Celery for async task processing
    "celery[redis]~=5.3.6",
    "django-celery-beat~=2.8.1",  # Updated for Django 5.1 compatibility
    # LDAP for Active Directory integration
    "ldap3~=2.9.1",
    # Encryption for sensitive fields
    "cryptography>=41.0.0",
    # Circuit breakers for resilience
    "pybreaker>=1.4.0",
    # Retry decorators with backoff
    "tenacity>=8.2.0",
    # Distributed tracing & observability
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "opentelemetry-instrumentation-django>=0.42b0",
    "opentelemetry-instrumentation-celery>=0.42b0",
    "opentelemetry-instrumentation-requests>=0.42b0",
    "opentelemetry-exporter-otlp>=0.42b0",
    # Metrics & monitoring
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
# Development Dependencies
dev = [
    # Testing
    "pytest~=8.3.5",
    "pytest-django~=4.8.0",
    "pytest-cov~=6.1.1",
    "pytest-mock~=3.12.0",
    # Code Quality
    "black~=25.1.0",
    "flake8~=7.2.0",
    "mypy~=1.8.0",
    "django-stubs~=5.1.0",  # Updated to match Django 5.1
    "isort~=5.13.2",
    # Development Tools
    "django-extensions~=3.2.3",
    "ipython~=8.20.0",
    # Pre-commit
    "pre-commit~=3.6.0",
]

# Production Dependencies
prod = [
    # WSGI Server
    "gunicorn~=21.2.0",
    # ASGI Server (for async support)
    "uvicorn[standard]~=0.30.0",
    # Monitoring
    "prometheus-client~=0.20.0",
    # Sentry (Error Tracking)
    "sentry-sdk~=1.40.0",
]

[tool.setuptools]
packages = ["apps"]

# Black Configuration (BuildWorks.AI Standard: line-length = 120)
[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort Configuration (BuildWorks.AI Standard: profile = black)
[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
known_first_party = ["apps", "config"]

# MyPy Configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_strict_optional = true
plugins = ["mypy_django_plugin.main"]
explicit_package_bases = true
mypy_path = "."

[tool.django-stubs]
django_settings_module = "config.settings.development"

[[tool.mypy.overrides]]
module = "apps.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "apps.*.models",
    "apps.connectors.*.models",
]
# Django models.Field don't have type annotations by default
# django-stubs plugin handles this but we also need fallback for non-plugin runs
disable_error_code = ["misc", "var-annotated", "attr-defined"]

# Models with _p5, _p5_5 etc suffixes
[[tool.mypy.overrides]]
module = [
    "apps.evidence_store.models_p5_5",
    "apps.evidence_store.models_p5",
    "apps.cab_workflow.models_p5",
    "apps.ai_agents.models",
]
disable_error_code = ["misc", "var-annotated", "attr-defined"]

[[tool.mypy.overrides]]
module = [
    "apps.*.views",
    "apps.*.services",
    "apps.*.serializers",
]
# DRF ViewSets and services often return Any from super() calls
# and use dynamic QuerySet operations and model attribute access
disable_error_code = ["no-any-return", "attr-defined"]

[[tool.mypy.overrides]]
module = [
    "apps.evidence_store.*",
    "apps.cab_workflow.*",
]
# These modules use forward-referenced models and factory patterns
# that mypy can't resolve correctly
disable_error_code = ["attr-defined", "no-any-return"]

[[tool.mypy.overrides]]
module = [
    "apps.integrations.services.*",
]
# Integration services interact with external APIs
disable_error_code = ["no-any-return", "arg-type"]

[[tool.mypy.overrides]]
module = [
    "apps.connectors.*.client",
]
# Connector clients use dynamic API responses and mixed-type dicts
disable_error_code = ["no-any-return", "assignment", "var-annotated"]

[[tool.mypy.overrides]]
module = [
    "apps.core.demo_data",
    "apps.core.resilient_http",
    "apps.core.structured_logging",
    "apps.core.views_health",
]
# Core utilities use dynamic patterns and circuit breaker library
disable_error_code = ["attr-defined", "operator", "arg-type", "no-any-return", "return"]

[[tool.mypy.overrides]]
module = [
    "apps.evidence_store.blast_radius_classifier",
]
# Blast radius uses indexed objects from JSON
disable_error_code = ["index"]

[[tool.mypy.overrides]]
module = [
    "apps.packaging_factory.services",
]
# Uses django.contrib.auth.get_user_model() pattern
disable_error_code = ["valid-type"]

[[tool.mypy.overrides]]
module = [
    "apps.ai_agents.agents.*",
]
# Agent implementations use dynamic model references
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
module = [
    "apps.authentication.dev_auth",
]
# Dev auth has intentional type coercion
disable_error_code = ["assignment"]

[[tool.mypy.overrides]]
module = [
    "apps.ai_agents.providers.*",
]
# Async generator return types differ from coroutine signatures in base class
disable_error_code = ["override"]

# Pytest Configuration
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = ["__pycache__", ".git", ".venv", "node_modules"]
testpaths = ["tests", "apps"]
markers = ["django_db: mark test as requiring database"]
addopts = [
    "--cov=apps",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=90",
]
